Separate out access token model and cache
